requirements.txt
django ==2.1.7
pillow ==5.4.1
